Refactorización de funciones y vistas para mejor maquetado de la aplicación
