+++ src/Screens/Characters/RickAndMorty.js
@@ -1,0 +1,13 @@
+import react from "react";
+import ListCharacters from "./components/ListCharacters";
+import style from '../../../assets/Styles/RickAndMortyStyles/RickAndMortyStyle'
+
+function RickAndMortyView(){
+  return(
+    <>
+      <ListCharacters />
+    </>
+  )
+}
+
+export default RickAndMortyView
