+++ src/Screens/Characters/components/ViewDetail.js
